requirements.txt
pillow
pymupdf
reportlab
